requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
requests==2.31.0
pymongo==4.6.3
psycopg2-binary==2.9.9